Add skeleton for blocks
